Refactor TableStyle enums, improve WriteCmColor logic
Refactored TableStyle by moving CellStyleClass and CellContentLayoutFlags enums to separate partial class files for better organization. Enhanced WriteCmColor in DwgMergedStreamWriter to handle true color, ByLayer, and color index cases according to DWG format. Updated CSUtilities submodule to a clean commit and removed obsolete comments and TODOs.
