+++ src/ACadSharp/Objects/TableStyle.CellStyleClass.cs
@@ -1,0 +1,11 @@
+﻿namespace ACadSharp.Objects;
+
+public partial class TableStyle
+{
+	public enum CellStyleClass
+	{
+		Data = 1,
+
+		Label = 2
+	}
+}
